Lint: rewrite import handling using stable APIs
This rewrite avoids using Kotlin descriptors, which are going away
soon in the migration to K2.

The main idea here is: we don't need to resolve the exact symbol
being imported, because the requested import is already given
as a raw string.

Historical note: the IDE does not have good support for direct
manipulation of import statements. Instead, the recommended solution
is to use fully qualified names in inserted code, and then apply a
"reference shortener" pass afterward. Unfortunately, this does not
work for Kotlin extension functions because those cannot be
expressed in a fully qualified manner (see related issue KTIJ-24422).
As a workaround, ReplaceStringQuickFix supports "explicit" imports.

Test: existing tests in ReplaceStringQuickFixTest
Bug: relates to b/268404355
Change-Id: Iea3bda7ec0fc0f573016e48418c40828e1d16ec2

GitOrigin-RevId: 3895f0ee258016a4261fff0940e4b22dd71f60c9