Add CPU_TRACE event listeners to cpu tracing methods
In today's Profiler, the start and stop tracing methods found in
CpuProfiler only register listeners for the TRACE_STATUS events.
The other event needed to be listened to is the CPU_TRACE event,
which contains actual data about the trace captured. However, in
the current Profiler, this CPU_TRACE event is just one of many events
constantly polled for in the background by the InProgressTraceHandler.
The InProgressTraceHandler does much more than just receive and handle
the CPU_TRACE event, while for the prospective Task-Based UX framework,
that is all that is needed. Thus, registering a transport event listener
accomplishes this for us, removing the need for background polling.

The CPU_TRACE event is also polled for by the StudioProfiler's update
method. Within this method is a call to the artifact fetchers, which in
turn make queries on the data base for CPU_TRACE events and many more
types of events. This type of polling is also uneccessary for the Task-Based UX framework. The reason polling in general is uneccessary for the Task-Based UX framework is because each task will clearly outline the command and corresponding event to be expected.

Thus, to prepare for the move from polling to event-precise listeners,
the CPU_TRACE listener has been added. This will allow us to reuse the
start and stop capture methods of CpuProfilerStage for any CPU-tracing
related tasks built in the future.

Bug: 271230809
Test: manual
Change-Id: I966494df8018f6114991c63f73b99485a7b72ffd

GitOrigin-RevId: e51e78425d48e04ee7ebcb9edf0ba2ed76a6ade3