Extract transport event listener registration to util
The creation and registration of transport event listeners
is a common pattern found in the CpuProfiler start and stop
tracing methods, as well as in the general codebase.

As these event listeners will play a bigger role in the
Task-Based framework, moving this functionality to a
helper will be useful in reducing redundant and verbose code.

Bug: 271230809
Test: existing
Change-Id: If81820053c2b18a2214d90d0755111a0bde3d278

GitOrigin-RevId: eaae89931bedb25e9fa3abe511435228239229f5