Implement new Gradle JDK validation to support #GRADLE_LOCAL_JAVA_HOME
The GradleJdkValidationManager adds more granularity for the invalid
gradle root JDK configuration helping to provide a proper error
message to users, better failures tracking but also a recovery mode.

For this reason the InvalidGradleJdkCause defines all the possible
cases for an invalid JDK configuration. In addition, this cause is
linked with the new GradleJdkException base class that makes easily
to define any jdk exception like InvalidGradleLocalJavaHomeException
that indicates when the provided java.home under
.gradle/config.properties points to invalid or corrupted path.

This new validation is used by the existing recovery system triggered
from AndroidGradleProjectStartupActivity.runActivity that now
supports multiple gradle roots, and basically is responsible to
potentially recover to safe state the project JDK configuration
for the user, when they open their project. In addition to that, the new
validation is also used on GradleJvmNotificationExtension.customize
responsible to override the IntelliJ notification in case of
ExternalSystemJdkException, that also support multiple gradle roots.

Bug: 214428183
Test: GradleSyncInvalidJdkConfigurationIntegrationTest
Test: GradleJvmNotificationExtensionTest
Change-Id: I23f36e83785a051dacbfdfb47a422bf03e75f9c4

GitOrigin-RevId: 898ab6073e763ead0a8186ed2e9dbfd6c297bde3