Mitigate the save gif issue
A bug was reported that the Save Gif feature is not working.
The issue could be related to this change - ag/20568658
Due to the change, modelChangedOnLayout won't be invoked to set isRenderCompleted to true although the animation is still rendering.

This CL mitigates the issue by removing .assertCompleted() to avoid an exception being throw.

Test: Tested locally
Fixes: 263935841
Change-Id: I5667174656ee1c8c9237af1d88ac64e3c5aaff76

GitOrigin-RevId: f9cf2ba41c279078f9520e87af9fc045ad443d98