Support addPlatformDependency with version catalogs.
platform dependency (like BOM) is now supported by version catalogs.
When RecipeExecutor.addPlatformDependency is called, the version and the
library entires are added to the toml file (for a new project or a
module where existing project already uses version catalog)
and the reference to the bom entry is added to the build script.

Bug: 262009615
Bug: 254914985
Test: DefaultRecipeExecutorWithGradleModelTest, TemplateTest
Change-Id: I1577dcd0c368f149ca898816a7a178777ce7d082

GitOrigin-RevId: d3341b2baa1217672dc651d62d0421ed0b2bb3b1