Keep compiled ids while updated ones are calculated
Consecustive renderings reset the compile ids independently and
in parallel to the rendering. This allows the situation where
the ids are reset but rendering is still accessing them. It
causes Resources$NotFoundException.

Here we reset the ids into a temporary storage first and then
instantly swap them with the current ids.

Fixes: 267733147
Test: ResourceIdManagerTest
Change-Id: I90f8bfecb86799aa2bb625fbe8ff94b2bd3da5e2

GitOrigin-RevId: 28ec908e627acf4c7473d2190e24420c2a463429