[Gradle Ps] Add a convenience interface
I think in principle this could just be calculated directly by the
client by reaching into the GradleAndroidModel#androidProject field,
but that seems unwise, and possibly it ought to be an aim to remove
that field.  In any case, provide a getter for the computation of
flavor dimension names and the names of their associated
productFlavors.

Also strengthen a related test, though only of test framework code.

Bug: 271412864
Test: none
Change-Id: I5eb5099b99796071893af9913cf35b05a842deed

GitOrigin-RevId: 5aa3f38ffc4c681229bc9debc494d1d99ab8f2f2