Remove unnecessary error log for material icons
This error log was meant to catch icons that were in autocomplete but
for which we didn't have something to display. It did so, uncovering a
missing category of icons (extended material icons). That category is
tracked with b/272127130. Given that there's a large set of icons
missing now, it no longer makes sense to have an error log that will
recur.

Fixes: 271241284
Test: ComposeCompletionContributorTest
Change-Id: I0787c0ebff1d7c82b5ae286d20216817d81ba6a0

GitOrigin-RevId: ec148927cf3972461783279bc09d0c64361eea7c