Ensure Compose autocompletion works for override
`ComposeInsertHandler` is intended to work with completion of Composable
function invocations. To identify these, we look to see if the lookup
element corresponds to a function marked with @Composable. This ends up
being the case both for function invocations and for "overrides"
statements where the super function has @Composable.

This change contains some minor refactoring; the key difference is the
new check for the parent element to be a `KtNameReferenceExpression`.

Note that b/271675885 is not completely fixed by this change. The
completion for an override function will still not have @Composable
automatically added, resulting in code that doesn't compile. The Kotlin
plugin doesn't allow this ability, and I've filed
https://youtrack.jetbrains.com/issue/KTIJ-24899 to track.

Bug: 271675885
Test: ComposeCompletionContributorTest
Change-Id: Ia900a7a60460473a77d7e3db5154471ffab9f92d

GitOrigin-RevId: d4ab8377eb9e88615a2ec52afa88b088dfac789c