Add split button for two build/make types
This change removes the build (hammer) icon and
introduces a split button with two options: build
selected module(s) and build project.

A new group action is introduced that holds both
actions, and the top-level group is driving the
module selection.

The new UI does not support split buttons,
so when it is enabled only
"Build selected modules" button is shown.

Bug: 272338755
Test: MakeTypeSelectionGroupActionTest and manual
Change-Id: Ic092a374b75b529ee62d5e65e4ba3a47ce4318c1

GitOrigin-RevId: fdd89d8e4cede4d3612cd86092f0b78c764294ed