Rename StudioProfilersView classes
In this change, StudioProfilersView is renamed to SessionProfilersView
and BaseStudioProfilersView is renamed to StudioProfilersView. This is
a separate change from ag/21917330 so it's obvious in that change what
functionality is being extracted into the base class. With this change
most of the codebase now refers to the abstract StudioProfilersView so
we can move on to doing a different implementation for a task-based UX
and have the rest of the codebase work with that out of the box.

Bug: 270211065
Test: updated existing tests
Change-Id: I8ced7d262b2c207dfbed5d01c8b29b7a1d4776f8

GitOrigin-RevId: a1b3ad5221a166fb615dae7f12f947209ba72b65