Save all files before opening the Image Asset Wizard
The Image Asset Wizard is using Java File objects to create the icons.
The content of those are not necessarily in sync with the content of
files in the IDE. This saves all files on opening the Image Asset Wizard
to ensure that file modifications from the IDE are taken into account
when previewing and creating the new icons.

Fixes: 230583012
Test: checked manually
Change-Id: I3b1bf3265f825add9c3a345802b5b4feb5f9d6ee

GitOrigin-RevId: 1f6799392cfab6019ad2660bd1950c425acc0095