Do not update the overlay until current refresh is complete
When refreshing the preview and using Fast Preview with multiple
elements, there is a chance that the compilation of a new change
completes while the previous refresh is on-going. This change avoids
refreshing the overlay while a refresh is currently being processed. It
also adds a method that helps to know when a refresh has completed.

This is not the ideal solution and additional work needs to happen in
this area. There are mainly two options:
- Pass an immutable overlay to the ModuleClassLoader at construction
  time so it can not be changed. This has other risks like the overlay
  being deleted while the load is happening.
- Cancel the previous on-going refresh.

Fixes: 273303505
Test: Existing tests pass
Change-Id: I46296c19b39a6f57a2fb88d2dcef9ee479ce47aa

GitOrigin-RevId: 24bdc6e8af6c4a866b1d27ec4ac675f82aff3d95