Make adt-ui-compose module libraries exports
In order to allow for dependent modules to be able to
build their own Compose Desktop UI/access Compose Desktop UI
apis, the Compose module libraries in the adt-ui-compose module
have been made exports. This is because changing module
libraries to be exports exposes the underlying apis
to dependent modules.

Bug: 272822954
Test: none
Change-Id: I4cefdf36f3367debf8e971c955d00d11710192fa

GitOrigin-RevId: aaa171b752b44635ab40728070e1413960bbe5c8