Add adt-ui-compose dep to profilers-ui module
In order to develop Compose Desktop UI specific to
the Android Studio Profiler, the adt-ui-compose module
has been added as a dependency of the profilers-ui module.

The adt-ui-compose module's prebuilt dependencies
needed to be added to the org.jetbrains.android plugin
libs. This was done in order to have access to the
Compose module in the profilers-ui module. Because of
this change, the expected files for studio test needed
to be updated.

Bug: 272822954
Test: none
Change-Id: Icc165db4bb9e7539ab616b294920b6d2e9e2d155

GitOrigin-RevId: a8a5b77f2ad61c949b6151533f336f9297dbe8de