Store previous module selection in Make modules action
This change store previously selected modules in the
"Make selected modules" action. This is to support
use cases when focus is removed from editor/project
view to e.g. tools windows. In these scenarios, the
module info is lost, and we'd fall back to selecting
the first application module. This is not great as
e.g. when there is a compilation error in lib.main,
the build output window is shown, it grabs focus,
and app.main becomes the selected module to build.
This change prevents this from happening.

Bug: 273563125
Test: MakeGradleModuleActionTest
Change-Id: I2f4924f183bc8ebb2fb17502bdefe31fa6b6ca43

GitOrigin-RevId: 443d298f660c07cbf55723ae5e041e5900ff62a2