Use KTS for the reverted UI tests
In b/271321123, qa smoke tests were made to use Groovy build script
to bypass the errors with KTS build script.
This CL is to use the KTS for those tests.

One of the primary factor of the reason it failed was that NpwControl
class needed to care about the KTS when replacing the local maven
repository.

Fixes: 273739793
Fixes: 271092042
Test: This is for tests
Change-Id: I3393760eac9babbdff94ec202addc8ef38d5185d

GitOrigin-RevId: a8decc461042c21f524993fd9c2b6255e575ca9a