[Dsl] Support parsing and writing packaging
AGP 8.0.0-beta02 (and, conveniently for computations in version space,
AGP 8.1.0-alpha01) introduced `packaging` as a synonym for the
`packagingOptions` block in `android`, deprecating the latter at the
same time.

Implement a packaging() method as part of the android() model, making
it a synonym for packagingOptions().  When choosing how to construct a
missing block, select "packaging" if the context's AGP version is
8.0.0-beta02 or newer.

Bug: 272056424
Test: existing, added
Change-Id: I72893b2885d3e3781ede805b5930c52919a67812

GitOrigin-RevId: dd17c936e5e6249a7f01dcde469323d58694ff10