Add buffer for power rail data
There has been an issue where power rail data is
being overwritten and thus not found when performing
a system trace. It was found that the buffer was
wrapping and overwriting the power rail data, thus
causing this problem.

To address this, a new buffer has been added with a
size of 4 mb. This buffer will be used (as of now)
exclusively by the 'android.power' data source. The value
of 4 mb is sufficient for longer and more data-intensive
tracing.

Bug: 254116554
Test: updated
Change-Id: I66b9727aa42e7f25ab690dd010d790d614582bda

GitOrigin-RevId: 533ef041799469957eb3fda6aca22ceb8408e608