Refactor related DaggerElement filtering
The current division of responsibilities between `IndexValue` and
`DaggerElement` is not quite correct. When `IndexValue` is resolving to
a `DaggerElement`, it shouldn't care about what type it's looking for;
rather, it should only be checking that the element it returns matches
what was stored in the index and is a valid Dagger definition.

`DaggerElement` then can then filter related elements by type as it
sees fit. This makes sense because it knows more about the type of
relationship between the two items, and can contain different type logic
dependending on the specifics of the relationship.

Bug: 272735207
Test: Included
Change-Id: I8b4786d81b5239e937a786ca353ab1edd84e68ad

GitOrigin-RevId: 961feed83ffb1a8110a8190e872de65407a8d1dd