Update colors of selected rows
In order to render properly in both light and dark themes, we need to
change the text color of widgets when a row becomes selected. (While
theme changes are mostly handled by JBColor, row selection color changes
must be implemented manually.)

This introduces the TablePresentationManager class for flexible and
mostly automatic propagation of color changes; it can easily be extended
as needed in the future.

Bug: 274151409
Test: Included
Change-Id: I0c384a742dc9c195571dcb44cd9a114d19b2a0ba

GitOrigin-RevId: 659c82a15b6d8266b4deed6f0eb87de027ab56b3