Publish separate jar for android model builders
Android studio IDE uses gradle tooling API facility to inject model
providers that compute and add extra models during build action
execution that runs as part of Gradle Sync. Currently, these extra model
providers are part of project-system-gradle-sync module. With this
change, we publish a separate jar file that include the targets that
includes the code that runs as part of build action in Gradle daemon
process. This prevents the PayloadSerializer to include android.jar in
the Gradle's runtime classpath and instead just include this new jar
with only specific relevant targets.

Bug: 243114022
Test: OpenProjectTest

Change-Id: I5c7d8fe4e683b29b2b12a2d4ad73217be3c15d29

GitOrigin-RevId: 76a95de6e501af333aba2f13cbced1058fb431b3