Refactor SafeArgs cache services to share infrastructure.
Combines the core parts of SafeArgsCacheModuleService (Java) and
KtDescriptorCacheModuleService (Kotlin K1) into a NavInfoFetcher with
accessors to handle retrieving the current state of Jetpack Navigation.
This is in preparation for creating another cache service for Kotlin K2.

In addition, creates an NavStatusCache that caches response data and
invalidates it on SafeArgs state change, abstracting that functionality
out of SafeArgsCacheModuleService. (KtDescriptorCacheModuleService,
despite the name, does no caching of its generated descriptors.)

Test: Treehugger
Fixes: 159950623
Change-Id: Ie679fdd73877486a643eaedf57d64100563a27ca

GitOrigin-RevId: 0f9ff761ea95895dc12acecfbf043d6f9f779f6e