Make the ColorBlindFilter be a property of ScreenViewProvider
Previously, when changing the color blind modes, the
SetColorBlindModeAction would replace the entire ScreenViewProvider
of the design surface, completely dropping the existing one and
setting a new color blind screen provider. But this is not ideal
as the previous ScreenViewProvider may use different layers in
different scenarios (Compose Preview, Layout Editor, Glance, Wear).

This change, moves the color blind mode to be a property in the
ScreenViewProvider, so that each tool is responsible for handling
(or ignoring) the ColorBlindMode changes. Also in this change all
existing ScreenViewProviders are modified to use the ColorBlindMode
property by default when creating a ScreenViewLayer.

Fixes: 274076939
Test: SetColorBlindModeActionTest
Change-Id: Ic542af82e3425a4cc64364a494e481a645c36e5c

GitOrigin-RevId: 37c03462d83292bf84f05e20d266bb06d5fa2482