Allow ResourceTypeCompletionContributor to recurse for Kotlin
ResourceTypeCompletionContributor looks for annotations on annotations
to find annotations that are Android typedefs (@IntDef, e.g.).
Unfortunately, it is not Kotlin-aware and as such does not handle
the case where the annotation is written in Kotlin (even if it is
being used in Java). This change allows the the contributor to include
values from Kotlin annotations as well.

Bug: 274021363
Test: Coming as part of a dependent change
Change-Id: If156507d367beac49581fc0db24fdb9ae31a4a5e

GitOrigin-RevId: a038ac19e3d98e726851e6d69209c9f297834cd2