Refactor DaggerElement Type and elementPsiType
Both `DaggerElement.Type` and `DaggerElement.elementPsiType` have become
obsolete as development has gone on, and this change removes both.

The `Type` enum is being replaced with concrete typing, since there's
effectively a 1:1 mapping between enum values and the classes that
extend `DaggerElement`.

`elementPsiType` was fairly generic when first put in, but it's become
clear that each `DaggerElement` ends up using it in a different way, and
there's no need for a shared definition. Having different definitions
allows implementations to change; for example, `Providers` can actually
supported providing multiple types, so this enables updating the
property for that type specifically to return a list.

Test: Included
Bug: 274176243
Change-Id: Ia56b868943f21b9382c14b22b0ff1294bf9f6c84

GitOrigin-RevId: bcf1e8f24a96613f22edb48fb6c617932c19c351