Upgrade/clean compose dependencies
The current prebuilt compose dependencies the adt-ui-compose
and compose-samples modules use are v1.1.1 for the compose-desktop
deps and v0.7.16 for skiko-awt. These older prebuilts not only
lack more modern features, but also have bugs that are fixed
in later versions. The compose-desktop deps have been upgraded
to v1.3.1 which is last stable release as of today. The skiko-awt
module has been upgraded to v0.7.55 which addresses a major bug
with swing-compose interop.

It was also noticed that there were two compose-desktop-foundation
modules libararies defined in adt-ui-compose module, but only the
latter one declared was needed. Thus, the first one was removed.

Bug: 272822954
Test: manual
Change-Id: Iae7f70b216d56ec9fbfc471e1679f68cf4e9ab51

GitOrigin-RevId: e138d19139e3bf0c0ccfe127085a5a94a191e148