Calculating subtree sized of component roots added
Now extended and local reports will contain information of number and
total size of objects reachable from each component root.

To implement it we store a set of parent component roots for each
object. This set is propagated during the heap traverse, and after that
when we propagate each individual object - we add it's size to subtrees
of all parent objects. In order to optimize memory used for root sets
storage we only store hashcodes of the sets in the per-object entries.

Bug: 274892404
Test: HeapAnalyzerTest#testComponentRootSubtree
Change-Id: I13c0b5bd82825cd31cb94c718dfc3a1c0c89bb1f

GitOrigin-RevId: 7cdb496327bf570265b27e0d6117d40a20e26c35