Remove SwappableProcessHandler
...and replace with DeviceAwareProcessHandler

SwappableProcessHandler is not used for AC and ACC anymore, plus it had
misleading name as it used not only for swap related handlers, but for
all. We also don't need to store executor.

Check for DeviceAwareProcessHandler using user data instead of instance
of, because there are process handlers we don't create, so we can't
always extend DeviceAwareProcessHandler

Fixes: 274103920
Test: AndroidProcessHandlerTest + run tests
Change-Id: Ie868421f59910e160bf3e8c7cee90ea17b547cbf

GitOrigin-RevId: 89bde88a6c52a5b5e781888682e13b2f062934b5