Only show relevant screenshot frame options from logcat
The list of screenshot frame options when taking a screenshot from
logcat always included the options:
"Watch Square", "Watch Round", "Generic Phone" and "Generic Tablet"
regardless of if the device was a round watch or square watch. This CL
filters out the options that don't match with the current device.

Bug: 235061954
Test: added DeviceArtScreenshotOptionsTest, DeviceScreenshotOptionsTest
Change-Id: I1586ca63d9726e4e43c4e466d739da0a4a53361d

GitOrigin-RevId: ae8b200d4d028cdd1f991642eb8886dac381b4b7