Allow benchmarker to work consistently with emulators
Something changed recently with the way the emulator handles key
events in that some keys are not sent to the emulator until the key
is released. This change adds a KEY_RELEASED event to the KEY_PRESSED
event for dispatched keys so that they are properly received by the
emulator.

Benchmarking requires a sequence of numbers and other characters to
be entered by Studio, followed by the ENTER key. However, the ENTER
key was not being sent to emulators due to this issue.

Fixes: 267504025
Test: DeviceAdapterTest.kt
Change-Id: Id72ef991edb85bad30cb57cd753337f5fbb7d168

GitOrigin-RevId: 1290d3c30407bf82067403ca3b1dfa208120ed0a