[Insights VCS] Add inlay diff view links to traces in the stacktrace console
A "diff link" inlay element will be attached to the end of the
file line number of a trace line, if the file jump
is available (i.e. the class/file symbol from the line is resolvable)
and the target file is within the current project scope.

When a diff view is brought up, on the left panel is
the historical source file from the "affected" commit,
on the right panel is the current source file.

For smarter annotation (i.e. annotation would be shown
at the updated place if there's any code change), it would
be included in a following change.

Bug: 279773903
Test: will add
Change-Id: Ia0072c5d08a6784792763efda6dc72d52a5d14da

GitOrigin-RevId: 67d6ecccc7444933156521360dfc0c51b3ea05f4