Add an "Explain This" extension and hook it up in Studio
First, this topic adds a new extension point, IssueExplainer, which
all the various Android plugins can depend on. The service allows
clients to request details on a specific error.

In the LogCat window, you can right click on log entries and
crashes and ask for an explanation. (For crashes, it will
identify the root cause and only ask about the 5 deepest
stackframes along with the logging tag.)

Bug: 274933635
Test: ExplainLogcatCrashActionTest.*
Change-Id: Id2d2693d33a42f156e6502f9c838d968b498a243

GitOrigin-RevId: b3d6f51e6fbda20d17586eb03766814805620809