Switch targetSdkVersion warning quickfixes over to assistant
The quickfixes for lint's "not using latest targetSdkVersion" warning,
and "edited targetSdkVersion in editor" warning, are both pointing
to the TargetSdkVersion Assistant. However, there were two other
lint checks which offered quickfixes to just replace the value with
the latest; this CL also replaces these warnings' quickfixes with
the assistant launcher instead.

Test: Manual
Bug: Not filed
Change-Id: Id9a5426b6698931ff721076c6a3453d05441eeed

GitOrigin-RevId: e6cf5a13e9c606a36bd3af6d160dccfc25600296