Change how Running Devices tabs are handled
The content of Running Devices tabs is disposed as soon as the tab
is not visible (either closed, not the main tab or when the tool
window is collapsed). For this reason we should not keep references
to tabs components.
This change instead keeps a refenrece to tab ids, and references
to only the selected tab's components.

Bug: 265150325
Test: LayoutInspectorManagerTest, RunningDevicesStateObserverTest, ToggleLayoutInspectorActionTest
Change-Id: Ia201dd4a6469efae1f6c7a2869812c82aaed62a9

GitOrigin-RevId: 46106f2a76ad84ec3429202361119509ebdb8644