[PSD] Less Extract Variable in Version Catalog projects
We already do not offer the Extract Variable button for dependencies
defined in Version Catalogs.  The Project perspective in the PSD also
shows the AGP Version, and offers Extract Variable functionality; this
causes an IDE crash in projects where the AGP version is already
defined as a variable in the Version Catalog, as the attempt to lift
it to the buildscript scope ends up failing to generate valid Toml
literals.

Determining whether the AGP version is defined in a Version Catalog is
more difficult than determining whether an individual dependency is
from a Version Catalog; the project's AGP version can be in an
arbitrary number of different places, managed in an arbitrary number
of different ways (buildSrc dependency, plugin alias, plugins Dsl,
pluginManagement Dsl in settings, ....).  Play it safe by not offering
the button if the project has any version catalog files in use.

Fixes: 272655387
Test: manual
Change-Id: Idace8b49b876d75bf5032526f308d68990124eeb

GitOrigin-RevId: a934a84aff645c2b87246290dd64f7e8fb05420f