Adjust callers of GoogleMavenRepository
The new interfaces in GoogleMavenRepository work in terms of concrete
entities (e.g. Component) rather than uncertain
ones (GradleCoordinate).  This propagates into the Android Studio core
code primarily in the GradleDependencyCompatibilityAnalyzer, which
uses GradleCoordinate to represent at least three different things,
and calls various bits of sdk-common and the project system; it
essentially acts as a bottleneck in making gradual changes to the
implementation.

We have to preserve the GradleDependencyCompatibilityAnalyzer
interface, at least for now, because it is effectively
exposed (GradleCoordinate and all) to the generic Module System.  This
also prevents us from including an alternative entry point, at least
with the same name, as the types involved are not distinct at the JVM
platform level, differing only in type arguments.

However, we can reimplement the internals to make the types of the
various objects floating around (Dependency / Component / Module,
RichVersion / VersionRange / Version) explicit and manifest, and also
include in this file explicitly the places where we are using the
KnownVersionStability concept from GradleCoordinate.

Because we are working with different objects, the String return value
in the event of a detected incompatibility may change (it will have an
explicit representation of the requested range, rather than an
implicit representation of it from the KnownVersionStability.  This
affects the third return value of the analysis in one test.

Bug: 242691473
Bug: 279886738
Test: existing tests adjusted
Change-Id: I609e45390c2e4877244335031cff9de62cfa4a37

GitOrigin-RevId: f4cc6a172b90d1c238ea32734a59da4b65d16d5e