Fix ResourceFolderRepository's handling of dir deletion
The way this was written before, the call to `iterator.remove()` was
unconditional. As such, if you deleted two directories, by the time
you got around to handling the second deletion, `mySources` would be
empty, and `removeResourcesContainedInFileOrDirectory` would be a noop,
leaving the resources that should have been deleted dangling.

Fixes: 274425678
Test: ResourceFolderRepositoryTest.java
Change-Id: Ia52e3153a9a1043e7c95facdd47d8678e7538b0f

GitOrigin-RevId: 106325cd6bbeabd9582c896ee91b7ead7ecdd4e2