Replace Edit button with Overflow button
The Overflow button will contain whatever additional actions are needed
for a device or template, such as Wear pairing, reservations, editing
AVDs, etc.

These will be implemented as standard AnActions, using the DataContext
to obtain the needed DeviceRowData or DeviceHandle from the current row.
EditButton's action handler is implemented as an AnAction.

Bug: 260752291
Test: TODO
Change-Id: I2f4780bf5442b4b1bfe9364ff8fe65a9d6ad33d6

GitOrigin-RevId: 0bb6ae5c52438341e4c9d440eeaeff4dc8c2df09