Navigate on click for single Dagger result
This brings the v2 implementation inline with the v1 implementation,
both in terms of navigation (skipping the popup dialog when there's only
a single result) and icon display (not showing an icon when there are no
results).

I'm still toying with always showing the icon, which would allow us to
defer querying for the related items unless the user actually wants to
see them. But if that proves necessary, I'll add some of this logic back
in a separate change with test updates at that time, to ensure it's
clear exactly what the impacts are.

Test: DaggerRelatedItemLineMarkerProviderTestV2
Bug: 272736137
Change-Id: Ibbe7ffb900d7516ecc4c75b134246c16610ed988

GitOrigin-RevId: da66589a60958beac3e47b5556bd291ff0bba788