Add a timer in InformationPopup when mouse out
Fixes a bug that closes unexpectedly "Build and Refresh" button.
A previous solution adopted was calculating the edges and paddings of
both the popup and the parent button.
However, there is still the case the setCancelOnMouseOutCallback closes
the popup right on the mouse out of the parent button, not giving enough
 time to reach the popup.

The solution is to start a Java Swing timer on the mouse out from the
parent button to prevent it from closing unexpectedly.
When the mouse is over the popup or the button area the timer stops,
on the mouse out, the timer restarts.

Test: InformationPopupImplTest
Fixes: 274687239
Change-Id: If9e90188c2e081fddb3977ffb05bdc1ea9af7d0e

GitOrigin-RevId: e3cc81e9b344c47da7dd26c754893c921c5769bd