Make PreviewLifecycleManager to only (de)activate when needed
Make sure to only activate/deactivate when it was not already
activated/deactivated, as that is the behaviour that the usages of
PreviewLifecycleManager expect. For example for Compose Preview,
activating an already active representation would mean that the
flows will be initialized again, causing a pair of refreshes to
execute concurrently for a brief time, what could cause problems,
like duplicated previews.

Fixes: 271455196
Test: manually verified
Change-Id: I11295ae4552d716853a42fed4566a6b7902b939b

GitOrigin-RevId: ba852dddcc4e5b98802c96da02059566a4132983