Display Kotlin sources under Kotlin node in the Android view
Make kotlin sourceType be mapped to both java/ and kotlin/ folders,
do this by allowing source roots intersection for java and kotlin
sourcetypes.

Add filter to java and kotlin nodes with the following logic.
For Java node we want to remove kotlin files but only if kotlin node
would also contain the source this file belongs. Otherwise the file is not shown anywhere and we want to avoid that.
For Kotlin node we want to show all the files under purely kotlin
sources (keep as it was before) but additionally show kotlin only files from java source folders.

Bug: 273905575
Test: manually verified, updated tests snapshots
Change-Id: I2b26f7c2878853aa77ffb7f31ab44f02392f394f

GitOrigin-RevId: 59531669fffe21cbb248074cd0d225a9aaab5090