Enable expansion in a property table
Expansion on hover was disabled in a properties table because:
- Intellij disabled expansion for tables outside scroll panes.
- We had never implemented expansion for the value column.

It turns out that a JTextField has some hard coded margin at the
right edge of the control. This margin will show up as a hole in
the expanded control. Most control types use a JTextField.
I selected to render the text as a JLabel instead to avoid this
hole.

An additional problem was the handling of error borders, which is
either a red or yellow border around all 4 sides of each renderer.
When an item is expanded, the item is painted in 2 parts:
- the part inside the table cell
- the part outside the table cell (the popup)
The part inside the table cell should not display the right edge
of the popup, since that would make the expanded text have a
colored border line in the middle of the text.

Another problem was that we dont want to expand a combo box when the
user hovers over the dropdown button. If we did the button would be
shown in the popup, and the popup goes away as soon as the user
moves the mouse outside the table cell i.e. it would be impossible
to click on the dropdown.

Added testing:
- added tests for detecting the need for expansion

Fixes: 276488570
Test: Added unit tests
Change-Id: If1fd2a0a0f9c3f9699f388e0fcd8a312f0f2db91

GitOrigin-RevId: a585fc6535c2f3817f6d8b074b307598a578ea8a