[PSD] Give our replacement Action some default properties
If we do not have some text, at least, then there are circumstances
when the menu might be being built at a point where the project is not
present, and hence we will not give the presentation any text.
Without any default text in the action's template presentation, this
will lead to an IDE crash from the ActionMenu system in the platform.

Since our action's properties are effectively unchanging anyway, we
might as well initialize the template presentation with them; then
`update()` has to do less work.  Also rewrite update() so that it
hides the Project Structure Action when the project is missing.

Bug: 277369057
Test: none
Change-Id: Ifcaa086f55f5aabdba5c378dd677ce064cd525d3

GitOrigin-RevId: 5e843bdf3437acd51d2f97b1356a102981b53266