[Vitals] Parse report text to structured stack trace groups
We are not giving perfect solutions here but just
best efforts. If a trace is parsable, we will be
able to collect file name, line number or so for
annotations as part of the App Insights (Vitals) support.

For printing traces in our console, concatenating
everything like we do for Crashlytics will still work with
this change, even we might have imperfect trace groups
extracted.

Bug: 265153845
Test: added
Change-Id: Ifbfc3f9b1947a4d6500e2bc6d25e5462fef5f0bb

GitOrigin-RevId: 2196626aba438330fd412c317139671efb7f1cb3