Remove remaining uses of GradleVersion in android core code
This isn't actually such a radical change as it sounds, given that the
replacement is also explicitly gradle-specific, albeit in its package
`com.android.ide.common.gradle`, so there is still work to do to
abstract away the build-system-specific nature of some of the
operations.  But this at least gives some reassurance over
implementing correct semantics for at least *one* build system.

Bug: 242691473
Bug: 272742663
Test: existing (no change in functionality)
Change-Id: Ife6593f05735ef8e6bbe1a76a42330702d4e01bb

GitOrigin-RevId: 6703561774d1dff6dcc47242a013c5c46b613689