Add power data processing to studio-side
Before, all data processing was done in the trace
processor daemon. Now, the trace processor daemon
only returns the raw/cumulative data for power.

This change implements the data processing in studio-side,
dependent on the flag value. The only special processing
that was needed to be implemented was for the "Delta"
display mode, as the other modes were either the raw/cumulative
data, or hiding the data.

The "minmax" display mode has also been renamed
to "cumulative", as it was functionally showing
the raw/cumulative values anyways.

Bug: 254116554
Test: added + updated
Change-Id: I9be924966b2274ea43308b5575adcbc828746689

GitOrigin-RevId: c050f5d28b913f1e838acf0c6d3da082db2014cf