Add aggregation logic for power rail counters
There are some counters that we want grouped, which
entails aggregating their series data and displaying it
all into one track. An example of a collection of counters
that are grouped is the ddr power rails. When present, ddr.a
and ddr.c rails will have their data aggregated and displayed
under one "Memory" track.

The aggregation method requires a mapping of the counters
to a respective group, meaning any type of counter can
use this logic. However, in this change, only a power
rail counter grouping map has been added.

Bug: 254116554
Test: added + updated
Change-Id: Ie699a006e3934f9d9cf79b6b0c779a9692307588

GitOrigin-RevId: 3f2a056cee21009b7cad15bfe8246e6b7d36382d