Add counter aggregation start time normalization
A prior change added functionality to aggregate multiple
counters' data. In some cases, when aggregating two or
more counters with data in different magnitudes, the
first few points have massive deltas/differences between
values. This can be misleading if one wants to display the
delta view of the data, as the first points can potentially
spike in value.

Thus, the ability to normalize the start time has been
added to the aggregation function. It can be enabled
via a boolean parameter.

Explanation of "start time normalization":
To normalize the start time across aggregated counters,
the maximum start time of all counters is taken. This
becomes the cut-off timestamp for all aggregated counters.
For example, if counter 1, 2, and 3 start at timestamp
100, 100, 102 respectively, the aggregated data wont start
until timestamp 102.

Bug: 254116554
Test: added + updated
Change-Id: Ibde2312ad84db167b0e558b0cca31110abad0a43

GitOrigin-RevId: 4c8613426e3eadd6047005dbe9c8e686665447f6