Specify type of alignment in autocomplete
When displaying autocomplete for Compose alignment values, we override
the default completions from the Kotlin plugin in some cases. In those
cases, the resulting lookup element is missing a type hint (eg,
"Alignment.Horizontal") in gray text to the right of the proposed
completion. This change adds in such a hint.

Bug: 267196388
Test: ComposeImplementationsCompletionContributorTest
Change-Id: I3b9d11698c7c2e7f620914433a9d3b09e1ece9c1

GitOrigin-RevId: cac4c69b48d15a66f1f4251928358f9cb4b861e7