Update Arrangement/Alignment contributor
This contributor makes suggestions for Compose Arrangement and Alignment
types. The current implementation handled Horizontal and Vertical
variants of these, but is missing handling for two-dimensional variants.
It also is missing suggestions for various absolute positioning classes.

This change adds support for the missing variants, and refactors the
code in preparation for adding a weigher that can control ordering of
the results. (The heavier logic is being moved out of the contributors
into utility classes that are more easily reused by the weigher.)

I've opted to make the supported types mostly declarative and simply
listed in the code. This is because the structure of Arrangement and
Alignment is quite similar, but different enough that I found more
generic code was littered with corner cases. Since the structure is
unlikely to change, and there still is dynamic code to look for the
actual suggestions in updated versions of the library, I found the
declarative implementation to be best here.

Test: ComposePositioningCompletionContributorTest
Bug: 267196388
Change-Id: Id9cca3aca2d16caf113d627e469650d1591342fc

GitOrigin-RevId: 7ebe92018d1e946bdcde9241d2d8acee9c86a8fd