Make  battery track group tooltip order-agnostic
The current battery track group tooltip relies on
the initial ordering of the battery tracks to explain
the meaning of each track. Because the user can
reorder these tracks, the tooltip info can become
misaligned. Thus, the track names have been called
out individually in the tooltip to prevent this
misalignment.

Fixes: 279083791
Bug: 254116554
Test: manual
Change-Id: I7749a9ba3dd45a96986b6d9db201e95c66306a56

GitOrigin-RevId: bb9bda3178669103b56c77eef9951b9df97ff859