Closing editors on module remove
If module name is changed, IntelliJ treats it as a module delete/add
opeartion. This causes the Layout Editor to end up in an inconsistent
state where the text editor works but the preview shows a message
indicating that the file does not exist anymore.
This does not help the user.

The change closes the editor that was attached to that module so the
user can re-open it.

Test: Manually verified
Fixes: 278845008
Change-Id: I5b3b6284cbbe2d9f9474211836ef7dbec46d49e0

GitOrigin-RevId: b2ad5b81b402a44fa8e4c27179c909bf34a25c69