Fix Gradle sync for old projects that are using JDK < 17
After the studio migration to JDK 17, was revealed a problem about
adding the full android.jar to Gradle daemon classpath that was
addressed by ag/22162851. As part of the tooling extensions, the
android-extensions-ide.jar is also included during the
creation of Gradle init script on GradleProjectResolver, containing
the following modules with different jvm_target:
 1) android-kotlin:intellij.android.kotlin.extensions
     - jvm_target = latest // 17
 2) android-kotlin:intellij.android.kotlin.extensions.common
     - jvm_target = 8

Given a project that is using a JDK lower than 17 this will cause the
below exception due that both are packaged on the same archive, that
end up being included on the Gradle daemon classpath during sync:
GradleException: Failed to create Jar file android-extensions-ide.jar
  Caused by: java.lang.IllegalArgumentException:
    - Unsupported class file major version 61

Unfortunately its impossible to change target for the first one given
to its extended dependencies like "intellij.android.common" and many
others that require explicitly specifying the jvm_target. Instead, we
are going to package it on android-kotlin.jar that isn't related with
tooling extensions and as a consequence will not be included on the
Gradle daemon classpath causing issues when sync is using lower JDK
version than the compiled one.

This is possible since the only required tooling extension class is
AndroidExtensionsProjectResolverExtension that is contained under
"intellij.android.kotlin.extensions.common" module.

A proper integration test has been added to validate this scenario,
that can only be reproducible when running studio from nonsource code

Bug: 278521775
Bug: 274016286
Test: OldProjectAppTest
Change-Id: I7dae63fd8a5fd6932b8ddb44055c44833ce6b1b4

GitOrigin-RevId: 0a45668405126d3f3c9029ff06344e162595b5b0