Render decorators after device frame is rendered
Change the rendering order of decoration painters, so that they
are painted after the device frame is rendered.

In the case of Layout Inspector, which for now is the only user
of this, the painting should happen after the device frame is
rendered, to avoid it being covered by the device frame.

Test: NA
Bug: 265150325
Change-Id: Ib0047f95af6da498162993bebc8ba31efcca4ce0

GitOrigin-RevId: 3f64ce8597d0d00cf43d6d21406762f93073a918