Create new graphics objects before applying transformations
By creating a new graphics object we are sure the trasnformations
are applied only to that graphics object.

The graphics object of a component affects how its children are
painted. In this case applying the scale to the original
graphics object caused the children of EmulatorView and DeviceView
to be painted with the scale applyed.

Test: NA
Bug: NA
Change-Id: I845ce7a9527c0b129250de68a14a6977b9ea5209

GitOrigin-RevId: 9e6ac324ef5bba5da291ae45758218a7d2f815e5