Show screenshot options combo when the emulator has no skin
When an emulator didn't have a skin definition, the image post processor
would be passed as null to the ScreenshotViewer. The ScreenshotViewer
would hide the options combo box in that case, meaning the "Play
Compatible" option would also be hidden.

This change changes the logic to hide the "Show Device Frame" option but
keep the combo box visible. It also changes the contract of the
ScreenshotViewer to make the post processor mandatory. The only place
where we provided a null post processor was when the skin definition of
an emulator was null. The post processor can handle that specific case
and doesn't need to be omitted completely.

Bug: 235061954
Bug: 277021497
Test: updated ScreenshotViewerTest and EmulatorScreenshotActionTest +
manual testing

Change-Id: I248787b8b24fc86c21de69dbeed50510c2e10070

GitOrigin-RevId: 81955f483dd0a9fe9c6fc1015599f047a01ecb44