Distinguish between any Kotlin file and Kotlin source
Our custom editor has special, additional behaviours for Kotlin
(and Java) files as part of Android Source.  However, Kotlin Script
build files also have a fileType of KotlinFileType.INSTANCE, and so
are potentially accepted by the custom editor provider.

Restrict isSourceFileType() to return true on Kotlin files with an
extension of the default Kotlin extension (only) as well as all Java
files in the project.

Bug: 278839073
Test: manual
Change-Id: Ie8ec69995288ee72e9de974b8c8468ddbc164c1c

GitOrigin-RevId: 78d9509f025ed0b13212144b27198c0a3ed68809